Addition of relation to joinTree to plan join, paren and route
